lifi/pyproject.toml
[project]
name = "lifi-mcp"
version = "0.1.0"
description = "Li.Fi cross-chain bridge and DEX aggregator as an MCP server -- powered by Invariant Protocol"
requires-python = ">=3.10"
dependencies = [
    "invariant-protocol @ git+https://github.com/jim-technologies/invariantprotocol.git#subdirectory=python",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
